requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
backoff==2.2.1
bcrypt==4.2.0
bidict==0.23.1
blinker==1.8.2
build==1.2.2.post1
cachelib==0.13.0
cachetools==5.5.2
certifi==2024.7.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
coloredlogs==15.0.1
crontab==1.0.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.6.1
durationpy==0.9
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
flash==1.0.3
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
flask-redis==0.4.0
Flask-RQ2==18.3
Flask-Session==0.8.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-SSE==1.0.0
Flask-WTF==1.2.1
flatbuffers==25.2.10
freezegun==1.5.1
fsspec==2025.3.0
google-auth==2.38.0
googleapis-common-protos==1.69.2
groq==0.9.0
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.4
kubernetes==32.0.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
numpy==2.0.2
oauthlib==3.2.2
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
posthog==3.23.0
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pysqlite3-binary==0.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
PyYAML==6.0.2
redis==5.0.8
regex==2024.7.24
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rq==1.16.2
rq-scheduler==0.13.1
rsa==4.9
shellingham==1.5.4
shortuuid==1.0.13
simple-websocket==1.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.46.1
sympy==1.13.3
tavily-python==0.3.5
tenacity==9.0.0
tiktoken==0.7.0
tokenizers==0.21.1
tomli==2.2.1
tqdm==4.66.4
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.1.2
zipp==3.19.2
